+++ Orchard/Services/ContainerService.swift
@@ -7,10 +7,8 @@ import Combine
 class ContainerService: ObservableObject {
     @Published var containers: [Container] = []
     @Published var images: [ContainerImage] = []
-    @Published var builders: [Builder] = []
     @Published var isLoading: Bool = false
     @Published var isImagesLoading: Bool = false
-    @Published var isBuildersLoading: Bool = false
     @Published var systemStatus: SystemStatus = .unknown
     @Published var systemStatusError: String?
     @Published var systemStatusVersionOverride: Bool = false
@@ -18,8 +16,6 @@ class ContainerService: ObservableObject {
     @Published var loadingContainers: Set<String> = []
     @Published var containerVersion: String?
     @Published var parsedContainerVersion: String?
-    @Published var isBuilderLoading = false
-    @Published var builderStatus: BuilderStatus = .stopped
     @Published var dnsDomains: [DNSDomain] = []
     @Published var isDNSLoading = false
     @Published var networks: [ContainerNetwork] = []
@@ -55,19 +51,27 @@ class ContainerService: ObservableObject {
     let settings: SettingsStore
     /// Opens container shells in the preferred terminal.
     let terminalLauncher: TerminalLauncher
+    /// BuildKit builder state and lifecycle.
+    let builderService: BuilderService
     private var cancellables = Set<AnyCancellable>()
 
     init(backend: ContainerBackend = LiveContainerBackend(), runner: CommandRunner = SystemCommandRunner()) {
         self.backend = backend
         self.runner = runner
+        let alertCenter = alertCenter
         let settings = SettingsStore(alertCenter: alertCenter)
         self.settings = settings
         self.terminalLauncher = TerminalLauncher(settings: settings, alertCenter: alertCenter)
-        // Re-publish the extracted store's changes so views observing this facade
+        let builderService = BuilderService(runner: runner, settings: settings, alertCenter: alertCenter)
+        self.builderService = builderService
+
+        // Re-publish the extracted stores' changes so views observing this facade
         // still update while the migration is in progress.
-        settings.objectWillChange
-            .sink { [weak self] in self?.objectWillChange.send() }
-            .store(in: &cancellables)
+        for store in [settings.objectWillChange, builderService.objectWillChange] {
+            store.sink { [weak self] in self?.objectWillChange.send() }.store(in: &cancellables)
+        }
+        // BuilderService needs to know whether the system is up (teardown guard).
+        builderService.systemIsRunning = { [weak self] in self?.systemStatus == .running }
     }
 
     // MARK: - Settings (forwarded to SettingsStore)
@@ -140,29 +144,6 @@ class ContainerService: ObservableObject {
                 return "version not yet supported"
             case .unsupportedVersion:
                 return "unsupported version"
-            }
-        }
-    }
-
-    enum BuilderStatus {
-        case stopped
-        case running
-
-        var color: Color {
-            switch self {
-            case .stopped:
-                return .gray
-            case .running:
-                return .green
-            }
-        }
-
-        var text: String {
-            switch self {
-            case .stopped:
-                return "Stopped"
-            case .running:
-                return "Running"
             }
         }
     }
@@ -243,80 +224,17 @@ class ContainerService: ObservableObject {
         }
     }
 
-    func loadBuilders() async {
-        await MainActor.run {
-            isBuildersLoading = true
-            self.alertCenter.dismiss()
-        }
+    // MARK: - Builders (forwarded to BuilderService)
 
-        var result: ProcessResult
-        do {
-            result = try await runner.run(
-                program: settings.safeContainerBinaryPath(),
-                arguments: ["builder", "status", "--format", "json"]
-            )
-        } catch {
-            result = ProcessResult(exitCode: -1, stdout: nil, stderr: error.localizedDescription)
-        }
+    var builders: [Builder] { builderService.builders }
+    var builderStatus: BuilderStatus { builderService.builderStatus }
+    var isBuilderLoading: Bool { builderService.isBuilderLoading }
+    var isBuildersLoading: Bool { builderService.isBuildersLoading }
 
-        if result.failed {
-            let detail = result.stderr?.trimmingCharacters(in: .whitespacesAndNewlines)
-            await MainActor.run {
-                self.builders = []
-                self.builderStatus = .stopped
-                self.isBuildersLoading = false
-                // Only surface to the user if the system is actually up. When it is
-                // stopped or being torn down, a failing background read is expected
-                // and NotRunningView already communicates the state — don't flash.
-                if self.systemStatus == .running {
-                    if let detail, !detail.isEmpty {
-                        self.alertCenter.error("Builder status could not be read: \(detail)")
-                    } else {
-                        self.alertCenter.error("Builder status could not be read (exit \(result.exitCode)).")
-                    }
-                }
-            }
-            if let detail, !detail.isEmpty {
-                Log.containers.error("Builder status command failed (exit \(result.exitCode)). Stderr:\n\(detail)")
-            } else {
-                Log.containers.error("Builder status command failed with unknown error (exit \(result.exitCode)).")
-            }
-            return
-        }
-
-        switch parseBuilderStatus(stdout: result.stdout ?? "") {
-        case .notRunning:
-            await MainActor.run {
-                self.builders = []
-                self.builderStatus = .stopped
-                self.isBuildersLoading = false
-            }
-            Log.containers.debug("Builder status indicates no builder present.")
-
-        case .builders(let list):
-            await MainActor.run {
-                withAnimation(.easeInOut(duration: 0.3)) {
-                    self.builders = list
-                }
-                self.builderStatus = (list.first?.status.lowercased() == "running") ? .running : .stopped
-                self.isBuildersLoading = false
-            }
-            for b in list {
-                Log.containers.debug("Builder: \(b.configuration.id), Status: \(b.status)")
-            }
-
-        case .decodeFailure(let preview):
-            Log.containers.error("Failed to decode builder status. Stdout preview (first 200 chars):\n\(preview)")
-            await MainActor.run {
-                self.builders = []
-                self.builderStatus = .stopped
-                self.isBuildersLoading = false
-                if self.systemStatus == .running {
-                    self.alertCenter.error("Builder status could not be read: unexpected response from the container service.")
-                }
-            }
-        }
-    }
+    func loadBuilders() async { await builderService.loadBuilders() }
+    func startBuilder() async { await builderService.startBuilder() }
+    func stopBuilder() async { await builderService.stopBuilder() }
+    func deleteBuilder() async { await builderService.deleteBuilder() }
 
     // MARK: - Container Stats Management
 
@@ -746,109 +664,6 @@ class ContainerService: ObservableObject {
         await MainActor.run {
             Log.containers.debug("Timeout reached for container \(id), removing loading state")
             loadingContainers.remove(id)
-        }
-    }
-
-    func startBuilder() async {
-        await MainActor.run {
-            isBuilderLoading = true
-            self.alertCenter.dismiss()
-        }
-
-        var result: ProcessResult
-        do {
-            result = try await runner.run(
-                program: settings.safeContainerBinaryPath(),
-                arguments: ["builder", "start"])
-
-            await MainActor.run {
-                if !result.failed {
-                    Log.containers.debug("Builder start command sent successfully")
-                    self.isBuilderLoading = false
-                    // Refresh builder status
-                    Task {
-                        await loadBuilders()
-                    }
-                } else {
-                    self.alertCenter.error("Failed to start builder: \(result.stderr ?? "Unknown error")")
-                    self.isBuilderLoading = false
-                }
-            }
-
-        } catch {
-            await MainActor.run {
-                self.isBuilderLoading = false
-                self.alertCenter.error("Failed to start builder: \(error.localizedDescription)")
-            }
-            Log.containers.error("Error starting builder: \(error.localizedDescription)")
-        }
-    }
-
-    func stopBuilder() async {
-        await MainActor.run {
-            isBuilderLoading = true
-            self.alertCenter.dismiss()
-        }
-
-        var result: ProcessResult
-        do {
-            result = try await runner.run(
-                program: settings.safeContainerBinaryPath(),
-                arguments: ["builder", "stop"])
-
-            await MainActor.run {
-                if !result.failed {
-                    Log.containers.debug("Builder stop command sent successfully")
-                    self.isBuilderLoading = false
-                    // Refresh builder status
-                    Task {
-                        await loadBuilders()
-                    }
-                } else {
-                    self.alertCenter.error("Failed to stop builder: \(result.stderr ?? "Unknown error")")
-                    self.isBuilderLoading = false
-                }
-            }
-
-        } catch {
-            await MainActor.run {
-                self.isBuilderLoading = false
-                self.alertCenter.error("Failed to stop builder: \(error.localizedDescription)")
-            }
-            Log.containers.error("Error stopping builder: \(error.localizedDescription)")
-        }
-    }
-
-    func deleteBuilder() async {
-        await MainActor.run {
-            isBuilderLoading = true
-            self.alertCenter.dismiss()
-        }
-
-        var result: ProcessResult
-        do {
-            result = try await runner.run(
-                program: settings.safeContainerBinaryPath(),
-                arguments: ["builder", "delete"])
-
-            await MainActor.run {
-                if !result.failed {
-                    Log.containers.debug("Builder delete command sent successfully")
-                    self.isBuilderLoading = false
-                    // Clear builders array since it was deleted
-                    self.builders = []
-                } else {
-                    self.alertCenter.error("Failed to delete builder: \(result.stderr ?? "Unknown error")")
-                    self.isBuilderLoading = false
-                }
-            }
-
-        } catch {
-            await MainActor.run {
-                self.isBuilderLoading = false
-                self.alertCenter.error("Failed to delete builder: \(error.localizedDescription)")
-            }
-            Log.containers.error("Error deleting builder: \(error.localizedDescription)")
         }
     }
